requirements.txt
aioscheduler==1.4.2
click==8.1.3
discord.py==2.0.1
praw==7.6.0
python-dotenv==0.21.0
requests==2.26.0
SQLAlchemy==1.4.41
